fix: Organize searchbox files similar to other compound components
