+++ src/components/index.ts
@@ -1,7 +1,6 @@
 export * from './swatch-bar'
 export * from './product-card';
 export * from './search-box';
-export * from './suggestions';
 export * from './theme';
 export * from './range';
 export * from './checkbox-group';
